backend/requirements.txt
Flask==3.0.0
requests==2.31.0
google-generativeai==0.3.0
Flask-Bootstrap==3.3.7.1